# openapi/Unload.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Unload.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/runs/history": {
      "get": {
        "tags": [
          "RunHistory"
        ],
        "operationId": "RunHistory_GetWorkflowHistory",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowHistory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowHistory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowHistory"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "tags": [
          "RunLaunch"
        ],
        "operationId": "RunLaunch_StartRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunAcceptedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "RunStatus"
        ],
        "operationId": "RunStatus_GetRuns",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunStatusInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunStatusInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunStatusInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/active": {
      "get": {
        "tags": [
          "RunStatus"
        ],
        "operationId": "RunStatus_GetActiveRun",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{correlationId}": {
      "get": {
        "tags": [
          "RunStatus"
        ],
        "operationId": "RunStatus_GetRunByCorrelationId",
        "parameters": [
          {
            "name": "correlationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemberRunLifecycleStatus": {
        "type": "integer"
      },
      "MemberRunStatusInfo": {
        "required": [
          "memberName",
          "status",
          "lastStep",
          "message",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "memberName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MemberRunLifecycleStatus"
          },
          "lastStep": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RunnerStep"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RunAcceptedResponse": {
        "required": [
          "correlationId",
          "runStatusPath",
          "hubPath",
          "subscribeMethod",
          "eventName",
          "runStatusEventName",
          "stopPath"
        ],
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "runStatusPath": {
            "type": "string"
          },
          "hubPath": {
            "type": "string"
          },
          "subscribeMethod": {
            "type": "string"
          },
          "eventName": {
            "type": "string"
          },
          "runStatusEventName": {
            "type": "string"
          },
          "stopPath": {
            "type": "string"
          }
        }
      },
...
      "RunLifecycleStatus": {
        "type": "integer"
      },
      "RunnerStep": {
        "type": "integer"
      },
      "RunOutputArtifactInfo": {
        "required": [
          "fileName",
          "filePath",
          "memberName",
          "scriptCode",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "memberName": {
            "type": [
              "null",
              "string"
            ]
          },
          "scriptCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RunStartRequest": {
        "required": [
          "memberCodes"
        ],
        "type": "object",
        "properties": {
          "memberCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetCodes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "adminOverride": {
            "type": "boolean",
            "default": false
          },
          "publishToGateway": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "RunStatusInfo": {
        "required": [
          "correlationId",
          "taskCode",
          "status",
          "targetCodes",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "taskCode": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunLifecycleStatus"
          },
          "targetCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastStep": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RunnerStep"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "outputPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "memberStatuses": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/MemberRunStatusInfo"
            }
          },
          "outputArtifacts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RunOutputArtifactInfo"
            }
          },
          "workerStatuses": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/RunWorkerStatusInfo"
            }
          },
          "senderBatches": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/SenderBatchStatusInfo"
            }
          },
          "publishToGateway": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "RunWorkerStatusInfo": {
        "required": [
          "workerId",
          "state",
          "scriptCode",
          "memberName",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "workerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "state": {
            "type": "string"
          },
          "scriptCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "memberName": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SenderBatchStatus": {
        "type": "integer"
      },
      "SenderBatchStatusInfo": {
        "required": [
          "batchId",
          "memberName",
          "status",
          "updatedAt",
          "sentFiles"
        ],
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string"
          },
          "memberName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SenderBatchStatus"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sentFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SenderFileDispatchStateInfo"
            }
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SenderFileDispatchStateInfo": {
        "required": [
          "filePath",
          "sentAt"
        ],
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TaskRecord": {
        "required": [
          "taskCode",
          "startedAt",
          "completedAt",
          "correlationId",
          "message",
          "scriptsExecuted",
          "filesWritten",
          "outputPath"
        ],
        "type": "object",
        "properties": {
          "taskCode": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "scriptsExecuted": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "filesWritten": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outputPath": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "WorkflowHistory": {
        "required": [
          "fromDayInclusive",
          "toDayInclusive",
          "days",
          "runs",
          "taskHistory"
        ],
        "type": "object",
        "properties": {
          "fromDayInclusive": {
            "type": "string",
            "format": "date"
          },
          "toDayInclusive": {
            "type": "string",
            "format": "date"
          },
          "days": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunStatusInfo"
            }
          },
          "taskHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRecord"
            }
          }
        }
      }
    }
  },
...
}