# web/webApp/src/app/generated/api/models/sender-batch-status-info.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SenderBatchStatus } from '../models/sender-batch-status';
import { SenderFileDispatchStateInfo } from '../models/sender-file-dispatch-state-info';
export interface SenderBatchStatusInfo {
  batchId: string;
  memberName: string;
  message?: (string | null);
  sentFiles: Array<SenderFileDispatchStateInfo>;
  status: SenderBatchStatus;
  updatedAt: string;
}