# web/webApp/src/app/generated/api/models/task-record.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskRecord {
  completedAt: string;
  correlationId: (string | null);
  filesWritten: (number | string | null);
  message: (string | null);
  outputPath: (string | null);
  scriptsExecuted: (number | string | null);
  startedAt: string;
  taskCode: string;
}